+++ packages/blocks/core/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Automated Tests:
 - Tests for the Spacer Block to ensure is supported by Blockera.
+- Added automated test for `Navigation Link Block` to check that the `Back to Parent` navigation buttons work correctly.
 
 ### Miscellaneous